now admins can completely manage categories of meal. then i have to complete leaving reviews logic

+++ handlers/users/admin_panel.py
@@ -21,12 +21,12 @@ async def get_all_users(message: types.Message):
             "/advert - Отправить рекламное сообщение\n\n"
             "/mcontrol - Открыть панель работы с позициями в меню\n\n"
             "/userlist - Выводит список пользователей \n"
-            "/adminlist - Выаодит список администраторов \n\n"
+            "/adminlist - Выводит список администраторов \n\n"
             "/adminadd - Добавить админа \n"
             "/adminremove - Разжаловать админа с должности \n\n"
             "/passwd - Задать пароль <b>самостоятельного</b> подключения к администраторскому доступу. ГОВОРИТЬ ЛИЧНО ТОЛЬКО СВОИМ В КОНФЕДЕНЦИАЛЬНЫХ ЧАТАХ, А ЛУЧШЕ ВЖИВУЮ!!!\n\n"
             "/imadmin - Самостоятельное получение администраторского доступа \n\n"
-            "/profile - Введите <b>Telegram ID</b> пользователя вместе с этой командой и получите его профиль, если позваляют найстроки конфеденциальности"
+            "/profile - Введите <b>Telegram ID</b> или <b>Username</b> пользователя вместе с этой командой и получите его профиль, если позваляют найстройки конфеденциальности"
         )
     else:
         await message.reply("Только для админов :D")
@@ -78,9 +78,15 @@ async def imadmin(message: types.message):
         config.admins.append(str(message.from_user.id))
         await message.answer("Ты теперь админ!")
         user = await db.select_user(telegram_id=message.from_user.id)
-        await notify_admins(f"Новый администратор - {user[1]}, @{user[2]}")
+        await notify_admins(
+            f"Новый администратор - <a href='tg://user?id={user[-2]}'>{user[1]}</a>, @{user[2]}\n\n"
+            "<b>Самостоятельный доступ!</b>"
+        )
     else:
-        print("Пароли не совпадают!")
+        await notify_admins(
+            f"Кто-то пытался стать админом!\n\n"
+            ""
+        )
 
 
 @dp.message_handler(commands='adminadd')
@@ -89,13 +95,19 @@ async def add_admin(message: types.Message):
         if message.get_args():
 
             telegram_id = int(message.get_args())
-            try:
-                config.admins.append(telegram_id)
-                await db.make_him_admin(telegram_id)
-                await message.reply("Теперь он новый администратор!")
-                await dp.bot.send_message(telegram_id, "Вам был предоставлен администорский доступ!")
-            except Exception as exception:
-                await message.reply(str(exception) + "\n" + "Ошибка возможно в том что вы неправильно ввели телеграм id, или такого пользователя в базе нет!")
+            if telegram_id not in config.admins:
+                try:
+                    await dp.bot.send_message(telegram_id, "Вам был предоставлен администорский доступ!")
+                    await db.make_him_admin(telegram_id)
+                    config.admins.append(telegram_id)
+                    await message.reply("Теперь он новый администратор!")
+                except Exception as exception:
+                    await message.reply(
+                        f"<b>{str(exception)}</b>\n\n"
+                        "Ошибка возможно в том что вы неправильно ввели телеграм id, или такого пользователя в базе нет!"
+                    )
+            else:
+                await message.reply("Он уже администратор!")
         else:
             await message.reply("Напииши телеграм id пользователя которого хочешь сделать админом вместе с командой!")
 
@@ -106,12 +118,17 @@ async def remove_admin(message: types.Message):
         if message.get_args():
 
             telegram_id = int(message.get_args())
-            try:
-                config.admins.remove(telegram_id)
-                await db.remove_admin(telegram_id)
-                await message.reply("Администратор разжалован с позиции!")
-                await dp.bot.send_message(telegram_id, "Вы были разжалованы с позиции администратора!")
-            except Exception as exception:
-                await message.reply(str(exception))
+            if telegram_id in config.admins:
+                try:
+                    await dp.bot.send_message(telegram_id, "Вы были разжалованы с позиции администратора!")
+                    await db.remove_admin(telegram_id)
+                    config.admins.remove(telegram_id)
+                    await message.reply("Администратор разжалован с позиции!")
+                except Exception as exception:
+                    await db.remove_admin(telegram_id)
+                    config.admins.remove(telegram_id)
+                    await message.reply(str(exception))
+            else:
+                await message.reply("Он не является администратором!")
         else:
             await message.reply("Напииши телеграм id админа которого хочешь разжаловать с должности вместе с командой!")
